app/requirements.txt
fastapi==0.106.0
uvicorn==0.24.0
docker==6.1.3
pydantic==2.5.2
psycopg2-binary==2.9.6